requirements.txt
keboola.component
keboola.utils
keboola.http-client
freezegun
mock
pydantic==2.10.1
duckdb==1.1.3
polars
pyarrow